# docs/swagger.yaml
definitions:
  app.ResponseModel:
    properties:
      data:
        type: object
      message:
        description: Code int         `json:"code"`
        type: string
    type: object
...
  models.AddRoleForm:
    properties:
      descs:
        type: string
      menu_access:
        items:
          $ref: '#/definitions/models.MenuAccessLevel1'
        type: array
      notes:
        type: string
      role_status:
        type: integer
    type: object
...
  models.EditRoleForm:
    properties:
      descs:
        type: string
      menu_access:
        items:
          $ref: '#/definitions/models.MenuAccessLevel1'
        type: array
      notes:
        type: string
      role_status:
        type: integer
      updated_at:
        type: string
      updated_by:
        type: string
    type: object
...
  models.MenuAccess:
    properties:
      is_read:
        type: boolean
      is_write:
        type: boolean
      level_no:
        type: integer
      menu_id:
        type: integer
      order_seq:
        type: integer
      parent_menu_id:
        type: integer
      title:
        type: string
    type: object
  models.MenuAccessLevel1:
    properties:
      is_read:
        type: boolean
      is_write:
        type: boolean
      level_2:
        items:
          $ref: '#/definitions/models.MenuAccessLevel2'
        type: array
      level_no:
        type: integer
      menu_id:
        type: integer
      order_seq:
        type: integer
      parent_menu_id:
        type: integer
      title:
        type: string
    type: object
  models.MenuAccessLevel2:
    properties:
      is_read:
        type: boolean
      is_write:
        type: boolean
      level_3:
        items:
          $ref: '#/definitions/models.MenuAccess'
        type: array
      level_no:
        type: integer
      menu_id:
        type: integer
      order_seq:
        type: integer
      parent_menu_id:
        type: integer
      title:
        type: string
    type: object
...
  models.ResponseModelList:
    properties:
      all_column:
        type: string
      data:
        type: object
      define_column:
        type: string
      define_size:
        type: string
      last_page:
        type: integer
      message:
        type: string
      page:
        type: integer
      total:
        type: integer
    type: object
...
info:
  contact:
    email: nuryantofattih@gmail.com
    name: Nuryanto
    url: https://www.linkedin.com/in/nuryanto-1b2721156/
  description: Backend REST API for golang starter
  license: {}
  title: Starter
  version: "1.0"
paths:
...
  /api/role:
    get:
      parameters:
      - description: Page
        in: query
        name: page
        required: true
        type: integer
      - description: PerPage
        in: query
        name: perpage
        required: true
        type: integer
      - description: Search
        in: query
        name: search
        type: string
      - description: InitSearch
        in: query
        name: initsearch
        type: string
      - description: SortField
        in: query
        name: sortfield
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ResponseModelList'
      security:
      - ApiKeyAuth: []
      summary: GetList SaRole
      tags:
      - Role
    post:
      parameters:
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.AddRoleForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Add Role
      tags:
      - Role
  /api/role/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Delete role
      tags:
      - Role
    get:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: GetById SaRole
      tags:
      - Role
    put:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      - description: 'req param #changes are possible to adjust the form of the registration
          form from frontend'
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/models.EditRoleForm'
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Update Role
      tags:
      - Role
  /api/role/menu_access:
    get:
      parameters:
      - description: RoleID
        in: query
        name: role_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/app.ResponseModel'
      security:
      - ApiKeyAuth: []
      summary: Get string Array Menu Access Role
      tags:
      - Role
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"